+++ source/components/Field.js
@@ -144,6 +144,42 @@ export const createFieldComponent: ComponentCreator = (dataFunctions: DataFuncti
     }
 
     componentDidUpdate(prevProps: ComponentProps) {
+      // Update touched property
+      if (this.props.touched !== prevProps.touched) {
+        this.context._reformRedux.field.setFieldTouched(this.props.name, this.props.touched);
+      }
+
+      // Update changed property
+      if (this.props.changed !== prevProps.changed) {
+        this.context._reformRedux.field.setFieldChanged(this.props.name, this.props.changed);
+      }
+
+      if (
+        ['radio', 'checkbox'].indexOf(this.props.type) !== -1 &&
+        this.context._reformRedux.form.fieldsCount[this.props.name] > 1 &&
+        this.props.checked !== prevProps.checked
+      ) {
+        // Dont change field value if it's was changed (checkboxes)
+        if (
+          this.props.checked &&
+          this.props.type === 'checkbox' &&
+          listIncludes(getIn(this.state.field, ['value']), this.props.value)
+        ) {
+          return;
+        }
+
+        // Dont change field value if it's was changed (radio)
+        if (
+          this.props.checked &&
+          this.props.type === 'radio' &&
+          is(getIn(this.state.field, ['value']), this.props.value)
+        ) {
+          return;
+        }
+
+        return this.changeFieldValue(this.getFieldValue(this.props.checked));
+      }
+
       // Update value only for single fields
       if (this.context._reformRedux.form.fieldsCount[this.props.name] > 1) {
         return;
@@ -182,7 +218,8 @@ export const createFieldComponent: ComponentCreator = (dataFunctions: DataFuncti
     };
 
     getFieldValue = (data: any): any => {
-      const isEvent = data.nativeEvent && data.nativeEvent instanceof Event;
+      const isEvent = data && data.nativeEvent && data.nativeEvent instanceof Event;
+
       if (this.isRadio()) {
         const checked = isEvent ? data.target.checked : data;
         return checked ? this.props.value : '';