+++ source/index.js
@@ -32,8 +32,12 @@ export {
   resetField,
   resetFields,
   removeField,
+  setFieldTouched,
+  setFieldsTouched,
+  setFieldChanged,
+  setFieldsChanged,
 } from './actions/Field';
-export { resetForm } from './actions/Form';
+export { resetForm, setFormSubmitted } from './actions/Form';
 
 // Components