+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.6.3] - 2018-11-4
+
+- Bugfix: Checkbox list after unmount & mount resets their value.
+
+## [1.6.2] - 2018-11-2
+
+- Set 0 fields count on field unmount.
+- Don't change value for list of checkboxes or radio if value already was changed.
+
+## [1.6.1] - 2018-10-31
+
+- Update value through props in list of checkbox or radio buttons.
+
+## [1.6.0] - 2018-10-30
+
+- New action `setFormSubmitted`.
+- Change touched and changed properties through props.
+
 ## [1.4.1] - 2018-10-22
 
 - Usage of polyfills.